added tests for runner (not yet working)
